test(lib-storage): remove duplicate file
